feat(provider/node): implement update function
The update function was missing, nothing
happen after updating the state of the node

Error: Provider produced inconsistent
result after apply

When applying changes to manta_node.node,
provider
"provider[\"registry.terraform.io/manta\"]"
produced an unexpected new value: .state:
was cty.StringVal("Off"), but now
cty.StringVal("On").

This is a bug in the provider, which should
be reported in the provider's own issue
tracker.